+++ microservices/product/test/product.test.js
@@ -1,0 +1,1 @@
+const app = require("./app");
